Fix a close bug (#48)
Previously, when a client closes PDC servers, it immediately returns without waiting for the server to acknowledge. If we try to start the server immediately after the close operation, restart checkpoints will fail. There is no way to check when the server checkpoint and cache write-back are finished or not. This pull request changes the behavior. The close_server program will not return until cache-write back and server checkpoint is finished. These two operations are processed before the mercury server looping is finished.